requirements.txt.txt
discord.py
instaloader

